[dist] Merge cli into main module
__main__.py is already in the main module so the CLI
logic should also be in the same location. We can
move shared logic later where it belongs.
