+++ setup.cfg
@@ -47,7 +47,6 @@ packages =
 # Main application
     dvr_scan
     dvr_scan.app
-    dvr_scan.cli
     dvr_scan.shared
 # Documentation (pre-built HTML but without images to keep size small)
     dvr_scan.docs
